backend/htx-backend/requirements.txt
annotated-types==0.7.0 
anyio==4.9.0 
certifi==2025.4.26 
click==8.2.1 
colorama==0.4.6 
eval-type-backport==0.2.2 
fastapi==0.115.12 
h11==0.16.0 
httpcore==1.0.9 
httpx==0.28.1 
idna==3.10 
mangum==0.19.0 
mistralai==1.8.1 
pydantic-core==2.33.2 
pydantic==2.11.5 
python-dateutil==2.9.0.post0 
python-dotenv==1.1.0 
six==1.17.0 
sniffio==1.3.1 
starlette==0.46.2 
typing-extensions==4.13.2 
typing-inspection==0.4.1 
uvicorn==0.34.3 
